Separated model subclasses in new modules
