+++ .gitignore
@@ -64,3 +64,6 @@ target/
 
 # Vim swapfiles
 .*.sw?
+
+# VSCode settings
+.vscode
